feat(rmdirfiles): Add function to clean a dir
Implement rmdirfiles function that will remove only the files in a
directory,
but not the directory itself.

Update package.json to build from source as a prepublish step and
remove the lib
directory from source control.

created by
space,
feature
libraries such as
created by
separated by a
 change to
body with the
longer